cb - refactored IdeaForm to seperate component
